Renderer now does not depends on the TreeView, only on DomConverter and Selection.
